SecurityComputing/requirements.txt
anyconfig==0.9.11
astroid==2.3.3
atomicwrites==1.4.0     
attrs==20.2.0      
certifi==2020.6.20      
cffi==1.14.3
chardet==3.0.4
click==7.1.2
clickclick==20.10.2     
colorama==0.4.4
connexion==2.7.0        
contextlib2==0.6.0.post1
cryptography==3.1.1     
Flask==1.1.2
Flask-Cors==3.0.9       
Flask-OpenTracing==1.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
idna==2.10
inflection==0.5.1
isort==4.3.21
itsdangerous==1.1.0
jaeger-client==4.3.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
mccabe==0.6.1
more-itertools==8.5.0
mysql-connector-python==8.0.22
openapi-spec-validator==0.2.9
opentracing==2.3.0
opentracing-instrumentation==3.3.1
packaging==20.4
pluggy==0.13.1
prometheus-client==0.8.0
protobuf==3.13.0
py==1.9.0
py-ms==2.4.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==5.4.1
python-dateutil==2.8.1
python-json-logger==0.1.11
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.15
swagger-ui-bundle==0.0.8
threadloop==1.0.2
thrift==0.13.0
tornado==5.1.1
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==1.0.1
geopy==2.1.0
